fix controllers with errors

+++ backend/app/controllers/forestController.ts
@@ -3,6 +3,7 @@ import BaseMapper from '../mappers/baseMapper';
 import { Forest } from '../types/index';
 import { AppError } from '../middlewares/errorHandler';
 import { catchAsync } from '../utils/catchAsync';
+import { forestSchema } from '../utils/shemasJoi';
 
 const forestMapper = new BaseMapper<Forest>('forest');
 
@@ -10,33 +11,63 @@ const forestController = {
 
     forests: catchAsync(async (req: Request, res: Response, next:NextFunction ): Promise<void>  => {
         const forests = await forestMapper.findAll();
-        res.json(forests);
+        if (forests.length === 0) {
+            res.json("No forests found");
+        }
+        res.status(200).json(forests);
     }),
-    forestById: catchAsync(async (req:Request, res:Response) => {
+    forestById: catchAsync(async (req:Request, res:Response, next: NextFunction) => {
         const id = parseInt(req.params.id, 10);
         const forest = await forestMapper.findById(id);
-        if (forest) {
-            res.json(forest);
-        } else {
-            res.status(404).send(`Forêt avec l'ID ${id} non trouvée`);
+        if (!forest) {
+            return next(new AppError(`Forest with ${id} not found`, 404));
         }
+        res.status(200).json(forest);
+
     }),
-    addForest: async (req:Request, res:Response) => {
-        const newForestData = req.body; 
+    addForest: catchAsync(async (req:Request, res:Response, next: NextFunction ) => {
+        const newForestData = req.body;
+        // Validation
+        const { error, value } = forestSchema.validate(newForestData);
+        if (error) {
+            return next(new AppError("Invalid data", 400));
+        }
+        // Forest exist
+        const existingForest = await forestMapper.findById(newForestData.id);
+        if (existingForest) {
+            return res.status(400).json({ message: "The forest already exists" });
+        }
+        // Create new forest
         const newForest = await forestMapper.create(newForestData);
         res.status(201).json(newForest);
-    },
-    updateForest: async (req:Request, res:Response) => {
+    }),
+    updateForest: catchAsync(async (req:Request, res:Response, next: NextFunction )  => {
         const id = parseInt(req.params.id, 10);
         const updatedForestData = req.body; 
+        // Validation
+        const { error, value } = forestSchema.validate(updatedForestData);
+        if (error) {
+            return next(new AppError("Invalid data", 400));
+        }
+        // Forest exist
+        const existingForest = await forestMapper.findById(id);
+        if (!existingForest) {
+            return res.status(404).json({ message: "Forest not found" });
+        }
+        // Update forest
         const updatedForest = await forestMapper.update(id, updatedForestData);
-        res.json(updatedForest);
-    }
-    ,
-    deleteForest: async (req:Request, res:Response) => {
+        res.status(200).json(updatedForest);
+    }),
+    deleteForest: catchAsync (async (req:Request, res:Response) => {
         const id = parseInt(req.params.id, 10);
+        // Forest exist
+        const existingForest = await forestMapper.findById(id);
+        if (!existingForest) {
+            return res.status(404).json({ message: "Forest not found" });
+        }
+        // Delete forest
         const deletedForest = await forestMapper.delete(id);
-        res.send("Forêt supprimée");
-    }
+        res.status(200).send("Forest deleted");
+    }),
 }
 export default forestController;
